GH-57: Filter capital scheme milestones by active
Note that the domain duplicates the milestone active values that exist in the
database. This is required for the in-memory repository and reference data that
are both used in tests.

This is a symptom of modelling milestone as a value object rather than an
entity. We need to address how we handle additional properties on value objects
like stage order and active.