GH-11: Update financials for a capital scheme
Introduce 'change financial' on domain that closes all current financials of the
same type. This works for financials such as 'spend to date' but not for ones
that can be adjusted like 'funding allocation'. Will raise a story to address
these financial types.

Repository can now update financials by using their surrogate ids. There is
currently an asymmetry in that we only load current financials, but the update
persists financials closed by the domain. Will refactor this in future to load
all financials, then introduce a get current financials repository method that
the get capital scheme endpoint can use.

Repository also needs to filter under embargo capital schemes. Currently one
could update an embargoed capital scheme's financials. Will raise a story.

Fake clock introduced so ITs can stub the current date and time. This is used to
set the effective date for created financials, and the closed date for
superseded financials.

The new endpoint unfortunately accesses the session directly to perform the
commit. A unit of work pattern needs to be introduced to abstract this away and
to make it testable in ITs. In the meantime, use aiosqlite to bridge SQLite to
async for ITs that exercise these routes to 'commit' the session. This can be
removed once we have a memory-based unit of work for ITs.

Finally, ITs don't assert that memory repositories were updated. We might be
able to improve on this in future by copying-on-read. E2E tests cover this in
the meantime.