GH-11: Add surrogate id to capital scheme financial
In order to update the list of financials for a capital scheme in the
persistence layer, we need to know their underlying database entity ids.

Although this constitutes as persistence leakage in the domain layer, it is an
acceptable compromise to otherwise decouple the layers.

In DDD this is known as a 'surrogate identity'.