Define foreign key with class attribute
Previously `CapitalSchemeEntity` referenced `CapitalSchemeFinancialEntity`
directly to define a relationship, so the foreign key had to be defined using
strings to avoid circular imports.

Now `CapitalSchemeFinancialEntity` has been extracted to a separate module and
that relationship is no longer defined, we can define the foreign key using a
class attribute instead.