GH-12: Extract aggregate root for capital scheme milestones
Milestones only need to be consistent with each other, so extract them to a
separate aggregate root that defines the boundary for transactional consistency.
This mirrors the work done for financials [1].

A new domain object for milestones allows the corresponding model object to
assume responsibility for marshalling to and from the domain.

We also allow the capital schemes in-memory repository to access this new
repository in order to filter by milestone.

[1] 431b4988aa090fdfc5d0d681b22d8ffea4aad515