GH-81: Fix financials repository to return none when not found
Previously we didn't distinguish between getting the financials for a capital
scheme that doesn't exist, and getting the empty set of financials for a capital
scheme; both where returned as an empty aggregate root.

Restructure the query to be anchored on capital schemes instead, then left join
to the (possibly empty) financials. This allows us to determine if the capital
scheme doesn't exist (empty result set), or the capital scheme has no
financials (the financial columns are null).