GH-73: Prohibit creating new funding allocations
Funding allocation is represented by a series of adjustments that when summed
provide the total amount. For this reason we should not close any funding
allocations. We need to consider how to change aggregated financials as part of
GH-74.

Prevent this in the domain by raising an error. Use custom validation [1][2] in
the model to avoid raising this error in the route and to return the correct 422
response.

[1] https://fastapi.tiangolo.com/tutorial/query-params-str-validations/#custom-validation
[2] https://docs.pydantic.dev/latest/concepts/validators/#model-validators